refactor: working on comments loading issue
