+++ src/components/comments/comments.jsx
@@ -1,3 +1,4 @@
+import { useState } from 'react';
 import { Button, Grid, Typography } from '@material-ui/core';
 import React, { useEffect } from 'react';
 import RefreshIcon from '@material-ui/icons/Refresh';
@@ -11,37 +12,7 @@ import { useSelector } from 'react-redux';
 import { loadComments } from '../../store/actions';
 import { useDispatch } from 'react-redux';
 import Loader from '../loader/loader';
-
-
-const buildTree = (comments, postId) => {
-  let commentsById = {};
-
-  comments.forEach((comment) => {
-    // console.log(`comment`, comment);
-    commentsById[comment.id] = comment;
-  });
-
-  
-  const rootComments = comments.filter((comment) => {
-    return comment?.parent?.toString() === postId;
-  });
-
-  return rootComments.map((item) => buildCommentTree(item, commentsById));
-};
-
-const buildCommentTree = (comment, commentsById) => {
-  const res = {
-    id: comment.id.toString(),
-    text: comment.text,
-    by: comment.by,
-  };
-  
-  res.children = comment.kids?.map((kidId) =>
-    buildCommentTree(commentsById[kidId], commentsById)
-  );
-  return res;
-};
-
+import { buildTree } from '../../utils/utils';
 
 const renderComments = (nodes) => {
   if (!nodes) return null;
@@ -56,7 +27,6 @@ const renderComments = (nodes) => {
             <div className={classes.commentWrap}>
               <p className={classes.commentBy}>{node.by}</p>
               <p
-                // fixme: it's not safe but come on, it's a test task for junior dev
                 dangerouslySetInnerHTML={{ __html: node.text }}
                 className={classes.comment}
               />
@@ -76,37 +46,46 @@ const renderComments = (nodes) => {
 function Comments(){
   const { id } = useParams();
   const dispatch = useDispatch();
+  const [loading, setLoading] = useState(false);
   const { comments } = useSelector((state) => state.newsItemCommentReducer);
   const singleComment = comments[id] || {};
   const { item, isLoaded } = singleComment;
-  
 
   useEffect(() => {
     dispatch(loadComments(id));
   }, []);
 
+  const handleRefreshComments = () => {
+    setLoading(true);
+    dispatch(loadComments(id));
+    setLoading(false);
+  };
+
+  console.log(`isLoaded`, isLoaded);
+
   if (!item) return null;
-  console.log(`item`, singleComment);
+ 
   const tree = buildTree(item, id);
   const commentsCount = item.length;
+ 
 
   return (
     <Grid item xs={12}>
-      <Grid container className={classes.boxWrapper}>
-        { 
-          isLoaded ? (
-            <>
-              <Grid item xs={10}>
-                <Typography variant="h5"> Comments { commentsCount } </Typography>
-              </Grid>
-              <Grid item xs={2}>
-                <Button
-                  variant="contained"
-                  color="primary"
-                  startIcon={<RefreshIcon />}
-                  onClick={() => {}}
-                > Refresh </Button>
-              </Grid>
+      <Grid container className={classes.boxWrapper}>    
+        <>
+          <Grid item xs={10}>
+            <Typography variant="h5"> Comments { commentsCount } </Typography>
+          </Grid>
+          <Grid item xs={2}>
+            <Button
+              variant="contained"
+              color="primary"
+              startIcon={<RefreshIcon />}
+              onClick={handleRefreshComments}
+            > Refresh </Button>
+          </Grid>
+          {
+            !loading ? (
               <Grid item xs={12}>
                 <TreeView
                   defaultCollapseIcon={<ExpandMoreIcon />}
@@ -117,9 +96,9 @@ function Comments(){
                 >
                   {renderComments(tree)}
                 </TreeView>
-              </Grid>
-            </>) : <Loader />          
-        }
+              </Grid>) :  <Loader />
+          }
+        </>
       </Grid>
     </Grid>
   );
