+++ src/store/newsItemCommentReducer.js
@@ -6,25 +6,19 @@ const initialState = {
 
 const newsItemCommentReducer = (state = initialState, action) => {
 
-  if (action.type === SETUP_COMMENT_ID){
-    
+  if (action.type === SETUP_COMMENT_ID){    
     const commentsObject = {};
     for (let i = 0; i < action.payload.length; i++) {
       const id = action.payload[i].id.toString();
       commentsObject[id] = { item: null, isLoaded: false };
     }
-
     return {
       ...state,
       comments: commentsObject
     };    
   }
 
-
   if (action.type === LOAD_COMMENT_SUCCESS) {
-
-    // console.log(`payload`, action.payload);
-
     return {
       comments:{
         ...state.comments,
@@ -35,8 +29,6 @@ const newsItemCommentReducer = (state = initialState, action) => {
       },
     };
   }
-
-
 
   return state;
 };
