+++ src/store/newsReducer.js
@@ -16,12 +16,10 @@ const newsReducer = (state = initialState, action) => {
   }
 
   if (action.type === LOAD_NEWS_SUCCESS) {
-
     const itemsObject = {};
     for (let i = 0; i < action.payload.length; i++) {
       itemsObject[action.payload[i].id.toString()] = action.payload[i];
     }
-
     return {
       ...state,
       loading: false,
